Move merge dictionaries and get supported names
